+++ include/utils/parse.js
@@ -2,11 +2,15 @@ const config = require(__dirname + '/config');
 
 module.exports = {
   getIp: function getIp(req) {
+    var ip = null;
     if (config.misc.cloudflareIP) {
-      return req.headers['cf-connecting-ip'] || req.headers['CF-Connecting-IP'] || req.headers['CF-CONNECTING-IP'] || req.ip;
+      ip = req.headers['cf-connecting-ip'] || req.headers['CF-Connecting-IP'] || req.headers['CF-CONNECTING-IP'];
+    }
+    if(!ip) {
+      ip = req.headers['x-forwarded-for'] || req.headers['X-Forwarded-For'] || req.headers['x-forwarded-host'] || req.headers['X-Forwarded-Host'] || req.ip;
     }
 
-    return req.ip;
+    return ip;
   },
   getDomainFromHeaders: (headers) => {
     var domain = 'default';